fix: moved cverainbowbadge from branchselector to branches page
